Migrate permissions models to the new app
