--- permissions/admin/__init__.py
@@ -1,7 +1,0 @@
-from .general_role import GeneralRoleAdmin
-from .unit_role import UnitRoleAdmin
-
-__all__ = [
-    "GeneralRoleAdmin",
-    "UnitRoleAdmin",
-]
