--- permissions/models/__init__.py
@@ -1,7 +1,0 @@
-from .general_role import GeneralRole
-from .unit_role import UnitRole
-
-__all__ = [
-    "GeneralRole",
-    "UnitRole",
-]
