+++ tests/test_graphql_api/test_permissions/test_permission_checker.py
@@ -3,8 +3,8 @@ from functools import partial
 import pytest
 from graphene_django_extensions.testing import build_query
 
-from permissions.enums import UserPermissionChoice, UserRoleChoice
 from tests.factories import UnitFactory, UnitGroupFactory, UnitRoleFactory, UserFactory
+from tilavarauspalvelu.enums import UserPermissionChoice, UserRoleChoice
 
 pytestmark = [
     pytest.mark.django_db,
