+++ tests/test_graphql_api/test_reservation/test_require_handling_permissions.py
@@ -1,7 +1,7 @@
 import pytest
 
-from permissions.enums import UserRoleChoice
 from tests.factories import ReservationFactory
+from tilavarauspalvelu.enums import UserRoleChoice
 
 from .helpers import REQUIRE_HANDLING_MUTATION, get_require_handling_data
