+++ tests/test_graphql_api/test_reservation/test_staff_modify_permissions.py
@@ -1,7 +1,7 @@
 import pytest
 
-from permissions.enums import UserRoleChoice
 from tests.factories import ReservationFactory
+from tilavarauspalvelu.enums import UserRoleChoice
 
 from .helpers import UPDATE_STAFF_MUTATION, get_staff_modify_data
