+++ tests/test_graphql_api/test_reservation/test_working_memo_update.py
@@ -1,7 +1,7 @@
 import pytest
 
-from permissions.enums import UserRoleChoice
 from tests.factories import ReservationFactory, ReservationUnitFactory, UserFactory
+from tilavarauspalvelu.enums import UserRoleChoice
 
 from .helpers import UPDATE_WORKING_MEMO_MUTATION, get_working_memo_update_data
