+++ tests/test_graphql_api/test_reservation_unit/test_update_permissions.py
@@ -1,7 +1,7 @@
 import pytest
 
-from permissions.enums import UserRoleChoice
 from tests.factories import ReservationUnitFactory
+from tilavarauspalvelu.enums import UserRoleChoice
 
 from .helpers import UPDATE_MUTATION, get_draft_update_input_data
