+++ tests/test_graphql_api/test_user/test_update_permissions.py
@@ -1,8 +1,8 @@
 import pytest
 from graphene_django_extensions.testing import build_mutation
 
-from permissions.enums import UserRoleChoice
 from tests.factories import UserFactory
+from tilavarauspalvelu.enums import UserRoleChoice
 
 # Applied to all tests
 pytestmark = [
