+++ tilavarauspalvelu/admin/__init__.py
@@ -1,3 +1,4 @@
+from .general_role.admin import GeneralRoleAdmin
 from .payment_accounting.admin import PaymentAccountingAdmin
 from .payment_merchant.admin import PaymentMerchantAdmin
 from .payment_order.admin import PaymentOrderAdmin
@@ -8,9 +9,11 @@ from .space.admin import SpaceAdmin
 from .terms_of_use.admin import TermsOfUseAdmin
 from .unit.admin import UnitAdmin
 from .unit_group.admin import UnitGroupAdmin
+from .unit_role.admin import UnitRoleAdmin
 from .user.admin import UserAdmin
 
 __all__ = [
+    "GeneralRoleAdmin",
     "PaymentAccountingAdmin",
     "PaymentMerchantAdmin",
     "PaymentOrderAdmin",
@@ -21,5 +24,6 @@ __all__ = [
     "TermsOfUseAdmin",
     "UnitAdmin",
     "UnitGroupAdmin",
+    "UnitRoleAdmin",
     "UserAdmin",
 ]
