+++ tilavarauspalvelu/admin/unit_role/admin.py
@@ -3,31 +3,37 @@ from django.db import models
 from django.utils.translation import gettext_lazy as _
 
 from common.typing import WSGIRequest
-from permissions.models import GeneralRole
+from tilavarauspalvelu.models import UnitRole
 
 __all__ = [
-    "GeneralRoleAdmin",
+    "UnitRoleAdmin",
 ]
 
 
-@admin.register(GeneralRole)
-class GeneralRoleAdmin(admin.ModelAdmin):
+@admin.register(UnitRole)
+class UnitRoleAdmin(admin.ModelAdmin):
     # Functions
     search_fields = [
         "user__username",
         "user__email",
         "user__first_name",
         "user__last_name",
+        "units__name",
+        "unit_groups__name",
     ]
-    search_help_text = _("Search by username, email, first name or last name")
+    search_help_text = _("Search by user's username, email, first name, last name, unit or unit group")
 
     # List
     list_display = [
         "role",
         "user_email",
+        "unit_names",
+        "unit_group_names",
     ]
     list_filter = [
         "role",
+        "units",
+        "unit_groups",
     ]
 
     # Form
@@ -37,15 +43,20 @@ class GeneralRoleAdmin(admin.ModelAdmin):
         "assigner",
         "created",
         "modified",
+        "units",
+        "unit_groups",
     ]
     readonly_fields = [
         "assigner",
         "created",
         "modified",
-        "assigner",
     ]
     autocomplete_fields = [
         "user",
+    ]
+    filter_horizontal = [
+        "units",
+        "unit_groups",
     ]
 
     def get_queryset(self, request: WSGIRequest) -> models.QuerySet:
@@ -56,12 +67,24 @@ class GeneralRoleAdmin(admin.ModelAdmin):
                 "user",
                 "assigner",
             )
+            .prefetch_related(
+                "units",
+                "unit_groups",
+            )
         )
 
-    def save_model(self, request: WSGIRequest, obj: GeneralRole, form, change: bool) -> GeneralRole:
+    def save_model(self, request: WSGIRequest, obj: UnitRole, form, change: bool) -> UnitRole:
         obj.assigner = request.user
         return super().save_model(request, obj, form, change)
 
     @admin.display(ordering="user__email")
-    def user_email(self, obj: GeneralRole) -> str:
+    def user_email(self, obj: UnitRole) -> str:
         return obj.user.email
+
+    @admin.display(ordering="units__name")
+    def unit_names(self, obj: UnitRole) -> str:
+        return ", ".join([unit.name for unit in obj.units.all()])
+
+    @admin.display(ordering="unit_groups__name")
+    def unit_group_names(self, obj: UnitRole) -> str:
+        return ", ".join([unit.name for unit in obj.unit_groups.all()])
