+++ tilavarauspalvelu/api/graphql/types/unit/filtersets.py
@@ -7,8 +7,8 @@ from graphene_django_extensions import ModelFilterSet
 from graphene_django_extensions.filters import IntMultipleChoiceFilter
 
 from common.date_utils import local_datetime
-from permissions.enums import UserRoleChoice
 from reservation_units.enums import ReservationKind
+from tilavarauspalvelu.enums import UserRoleChoice
 from tilavarauspalvelu.models import Unit
 from tilavarauspalvelu.models.unit.queryset import UnitQuerySet
