+++ tilavarauspalvelu/models/__init__.py
@@ -1,4 +1,5 @@
 from .building.model import Building
+from .general_role.model import GeneralRole
 from .location.model import Location
 from .payment_accounting.model import PaymentAccounting
 from .payment_merchant.model import PaymentMerchant
@@ -13,10 +14,12 @@ from .space.model import Space
 from .terms_of_use.model import TermsOfUse
 from .unit.model import Unit
 from .unit_group.model import UnitGroup
+from .unit_role.model import UnitRole
 from .user.model import ProfileUser, User
 
 __all__ = [
     "Building",
+    "GeneralRole",
     "Location",
     "PaymentAccounting",
     "PaymentMerchant",
@@ -32,5 +35,6 @@ __all__ = [
     "TermsOfUse",
     "Unit",
     "UnitGroup",
+    "UnitRole",
     "User",
 ]
