+++ tilavarauspalvelu/models/general_role/model.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from functools import cached_property
 from typing import TYPE_CHECKING
 
 from django.db import models
@@ -7,12 +8,17 @@ from django.utils.translation import gettext_lazy as _
 from graphene_django_extensions.fields.model import StrChoiceField
 
 from config.utils.auditlog_util import AuditLogger
-from permissions.enums import UserRoleChoice
+from tilavarauspalvelu.enums import UserRoleChoice
+
+from .queryset import GeneralRoleManager
 
 if TYPE_CHECKING:
     import datetime
 
     from tilavarauspalvelu.models import User
+
+    from .actions import GeneralRoleActions
+
 
 __all__ = [
     "GeneralRole",
@@ -29,6 +35,8 @@ class GeneralRole(models.Model):
     created: datetime.datetime = models.DateTimeField(auto_now_add=True)
     modified: datetime.datetime = models.DateTimeField(auto_now=True)
 
+    objects = GeneralRoleManager()
+
     class Meta:
         db_table = "general_role"
         base_manager_name = "objects"
@@ -38,6 +46,14 @@ class GeneralRole(models.Model):
 
     def __str__(self) -> str:
         return f"General Role '{self.role}' for {self.user.first_name} {self.user.last_name} ({self.user.email})"
+
+    @cached_property
+    def actions(self) -> GeneralRoleActions:
+        # Import actions inline to defer loading them.
+        # This allows us to avoid circular imports.
+        from .actions import GeneralRoleActions
+
+        return GeneralRoleActions(self)
 
 
 AuditLogger.register(GeneralRole)
