+++ tilavarauspalvelu/models/general_role/queryset.py
@@ -1,0 +1,14 @@
+from __future__ import annotations
+
+from django.db import models
+
+__all__ = [
+    "GeneralRoleManager",
+    "GeneralRoleQuerySet",
+]
+
+
+class GeneralRoleQuerySet(models.QuerySet): ...
+
+
+class GeneralRoleManager(models.Manager.from_queryset(GeneralRoleQuerySet)): ...
