+++ tilavarauspalvelu/models/unit_role/model.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from functools import cached_property
 from typing import TYPE_CHECKING
 
 from django.db import models
@@ -7,13 +8,16 @@ from django.utils.translation import gettext_lazy as _
 from graphene_django_extensions.fields.model import StrChoiceField
 
 from config.utils.auditlog_util import AuditLogger
-from permissions.enums import UserRoleChoice
-from permissions.models import GeneralRole
+from tilavarauspalvelu.enums import UserRoleChoice
+
+from .queryset import UnitRoleManager
 
 if TYPE_CHECKING:
     import datetime
 
     from tilavarauspalvelu.models import User
+
+    from .actions import UnitRoleActions
 
 __all__ = [
     "UnitRole",
@@ -33,6 +37,8 @@ class UnitRole(models.Model):
     created: datetime.datetime = models.DateTimeField(auto_now_add=True)
     modified: datetime.datetime = models.DateTimeField(auto_now=True)
 
+    objects = UnitRoleManager()
+
     class Meta:
         db_table = "unit_role"
         base_manager_name = "objects"
@@ -43,5 +49,13 @@ class UnitRole(models.Model):
     def __str__(self) -> str:
         return f"Unit Role '{self.role}' for {self.user.first_name} {self.user.last_name} ({self.user.email})"
 
+    @cached_property
+    def actions(self) -> UnitRoleActions:
+        # Import actions inline to defer loading them.
+        # This allows us to avoid circular imports.
+        from .actions import UnitRoleActions
 
-AuditLogger.register(GeneralRole)
+        return UnitRoleActions(self)
+
+
+AuditLogger.register(UnitRole)
