+++ tilavarauspalvelu/models/unit_role/queryset.py
@@ -1,0 +1,14 @@
+from __future__ import annotations
+
+from django.db import models
+
+__all__ = [
+    "UnitRoleManager",
+    "UnitRoleQuerySet",
+]
+
+
+class UnitRoleQuerySet(models.QuerySet): ...
+
+
+class UnitRoleManager(models.Manager.from_queryset(UnitRoleQuerySet)): ...
