+++ tilavarauspalvelu/typing.py
@@ -1,0 +1,4 @@
+from __future__ import annotations
+
+
+class permission(classmethod): ...  # noqa: N801
