+++ tilavarauspalvelu/utils/anonymisation.py
@@ -8,11 +8,10 @@ from social_django.models import UserSocialAuth
 
 from applications.enums import ApplicationStatusChoice
 from applications.models import Address, Application, ApplicationSection, Person
-from permissions.models import GeneralRole, UnitRole
 from reservations.enums import ReservationStateChoice, ReservationTypeChoice
 from reservations.models import Reservation
 from tilavarauspalvelu.enums import OrderStatus, ReservationNotification
-from tilavarauspalvelu.models import User
+from tilavarauspalvelu.models import GeneralRole, UnitRole, User
 
 ANONYMIZED = "Anonymized"
 SENSITIVE_RESERVATION = "Sensitive data of this reservation has been anonymized by a script"
