+++ tilavarauspalvelu/utils/helauth/pipeline.py
@@ -149,8 +149,8 @@ def migrate_from_tunnistamo_to_keycloak(*, email: str) -> None:
     new_user.save()
 
     from applications.models import Application
-    from permissions.models import GeneralRole, UnitRole
     from reservations.models import RecurringReservation, Reservation
+    from tilavarauspalvelu.models import GeneralRole, UnitRole
 
     # Migrate general roles.
     GeneralRole.objects.filter(user=old_user).update(user=new_user)
